+++ CHANGES.rst
@@ -8,6 +8,7 @@ Here you can see the full list of changes between each SQLAlchemy-Utils release.
 ^^^^^^^^^^^^^^^^^^^
 
 - Added Asterisk compiler
+- Added row_to_json FunctionElement
 - Made quote function accept dialect object as the first paremeter
 - Made has_index work with tables without primary keys (#148)